optimize metrics
