+++ common/src/main/java/org/tron/common/parameter/CommonParameter.java
@@ -483,6 +483,10 @@ public class CommonParameter {
   @Setter
   public static boolean ENERGY_LIMIT_HARD_FORK = false;
 
+  @Getter
+  @Setter
+  public boolean nodeMetricsEnable = false;
+
   private static double calcMaxTimeRatio() {
     //return max(2.0, min(5.0, 5 * 4.0 / max(Runtime.getRuntime().availableProcessors(), 1)));
     return 5.0;
