+++ common/src/main/java/org/tron/core/Constant.java
@@ -239,4 +239,6 @@ public class Constant {
   public static final String RATE_LIMITER_RPC = "rate.limiter.rpc";
 
   public static final String SEED_NODE_IP_LIST = "seed.node.ip.list";
+
+  public static final String NODE_METRICS_ENABLE = "metrics.nodeMetricsEnable";
 }
