+++ framework/src/main/java/org/tron/common/overlay/discover/node/NodeManager.java
@@ -34,7 +34,7 @@ import org.tron.common.utils.JsonUtil;
 import org.tron.core.ChainBaseManager;
 import org.tron.core.capsule.BytesCapsule;
 import org.tron.core.config.args.Args;
-import org.tron.core.metrics.MonitorMetric;
+import org.tron.core.metrics.MetricsService;
 
 @Slf4j(topic = "discover")
 @Component
@@ -63,7 +63,7 @@ public class NodeManager implements EventHandler {
   private ScheduledExecutorService pongTimer;
 
   @Autowired
-  MonitorMetric monitorMetric;
+  MetricsService metricsService;
 
   @Autowired
   public NodeManager(ChainBaseManager chainBaseManager) {
@@ -235,8 +235,8 @@ public class NodeManager implements EventHandler {
 
     NodeHandler nodeHandler = getNodeHandler(n);
     nodeHandler.getNodeStatistics().messageStatistics.addUdpInMessage(m.getType());
-    monitorMetric.getMeter(MonitorMetric.NET_UDP_IN_TRAFFIC)
-            .mark(udpEvent.getMessage().getData().length + 1);
+    metricsService.meterMark(MetricsService.NET_UDP_IN_TRAFFIC,
+            udpEvent.getMessage().getData().length + 1);
 
     switch (m.getType()) {
       case DISCOVER_PING:
@@ -259,8 +259,8 @@ public class NodeManager implements EventHandler {
   public void sendOutbound(UdpEvent udpEvent) {
     if (discoveryEnabled && messageSender != null) {
       messageSender.accept(udpEvent);
-      monitorMetric.getMeter(MonitorMetric.NET_UDP_OUT_TRAFFIC)
-              .mark(udpEvent.getMessage().getSendData().length);
+      metricsService.meterMark(MetricsService.NET_UDP_OUT_TRAFFIC,
+              udpEvent.getMessage().getSendData().length);
     }
   }
