+++ framework/src/main/java/org/tron/core/services/NodeInfoService.java
@@ -63,10 +63,8 @@ public class NodeInfoService {
     setConfigNodeInfo(nodeInfo);
     setBlockInfo(nodeInfo);
     setCheatWitnessInfo(nodeInfo);
-
     return nodeInfo;
   }
-
 
   private void setMachineInfo(NodeInfo nodeInfo) {
     MachineInfo machineInfo = new MachineInfo();
@@ -198,7 +196,6 @@ public class NodeInfoService {
     nodeInfo.setBeginSyncNum(dbManager.getSyncBeginNumber());
     nodeInfo.setBlock(chainBaseManager.getHeadBlockId().getString());
     nodeInfo.setSolidityBlock(chainBaseManager.getSolidBlockId().getString());
-
   }
 
   protected void setCheatWitnessInfo(NodeInfo nodeInfo) {
