+++ framework/src/main/java/org/tron/program/FullNode.java
@@ -76,7 +76,9 @@ public class FullNode {
 
     // http api server
     FullNodeHttpApiService httpApiService = context.getBean(FullNodeHttpApiService.class);
-    appT.addService(httpApiService);
+    if (CommonParameter.getInstance().fullNodeHttpEnable) {
+      appT.addService(httpApiService);
+    }
 
     // full node and solidity node fuse together
     // provide solidity rpc and http server on the full node.
@@ -86,7 +88,9 @@ public class FullNode {
       appT.addService(rpcApiServiceOnSolidity);
       HttpApiOnSolidityService httpApiOnSolidityService = context
           .getBean(HttpApiOnSolidityService.class);
-      appT.addService(httpApiOnSolidityService);
+      if (CommonParameter.getInstance().solidityNodeHttpEnable) {
+        appT.addService(httpApiOnSolidityService);
+      }
     }
 
     appT.initServices(parameter);
