+++ framework/src/main/resources/config.conf
@@ -564,3 +564,7 @@ event.subscribe = {
   }
 
 }
+
+metrics = {
+  nodeMetricsEnable = true
+}
